added created modals files and restructure
